% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_loglik_3.2.R
\name{bd_loglik_3.2}
\alias{bd_loglik_3.2}
\title{Loglikelihood for diversity-independent diversification model}
\usage{
bd_loglik_3.2(pars1, pars2, brts, missnumspec, methode = "lsoda")
}
\arguments{
\item{pars1}{numerical parameters of the model}

\item{pars2}{setting parameters of the model}

\item{brts}{A set of branching times, all positive.}

\item{missnumspec}{The number of species that are in the clade but missing in the phylogeny}

\item{methode}{The method used to solve the master equation, default is 'lsoda'.}
}
\value{
the loglikelihood value
}
\description{
Copy of the \code{bd_loglik()} function from DDD.3.2. For syntax and details, see \code{?bd_loglik()}.
}
\references{
- Etienne, R.S. et al. 2012, Proc. Roy. Soc. B 279: 1300-1309, doi: 10.1098/rspb.2011.1439

- Etienne, R.S. & B. Haegeman 2012. Am. Nat. 180: E75-E89, doi: 10.1086/667574
}
\author{
Rampal S. Etienne, Bart Haegeman & Cesar Martinez
}
