% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_loglik_choosepar_3.2.R
\name{bd_loglik_choosepar_3.2}
\alias{bd_loglik_choosepar_3.2}
\title{Internal DDD function}
\usage{
bd_loglik_choosepar_3.2(trparsopt, trparsfix, idparsopt, idparsfix, pars2,
  brts, missnumspec, methode)
}
\description{
Calls bd_loglik for bd_ML. Only difference with the function from DDD is that this one calls to bd_loglik_3.2.
}
\author{
Rampal S. Etienne & Bart Haegeman
}
