% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_ML_3.2.R
\name{bd_ML_3.2}
\alias{bd_ML_3.2}
\title{Maximization of the loglikelihood under the diversity-independent, possibly time-dependent diversification model (version 3.2)}
\usage{
bd_ML_3.2(brts, initparsopt = c(0.1, 0.05 * (tdmodel <= 1) + 10 *
  (length(brts) + missnumspec) * (tdmodel > 1)), idparsopt = c(1, 2 +
  (tdmodel > 1)), idparsfix = (1:4)[-idparsopt], parsfix = rep(0,
  4)[idparsfix], missnumspec = 0, tdmodel = 0, cond = 1,
  btorph = 1, soc = 2, tol = c(0.001, 1e-04, 1e-06), maxiter = 1000
  * round((1.25)^length(idparsopt)), changeloglikifnoconv = FALSE,
  optimmethod = "subplex", methode = "lsoda", verbose = 0)
}
\description{
Identical to DDD::bd_ML(), except that this one calls bd_loglik_3.2(). See DDD documentation for details.
}
\author{
Rampal S. Etienne & Bart Haegeman
}
